refactoring and proteksi api tag

+++ app/tag/controller.js
@@ -1,8 +1,16 @@
 const Tag = require("./model");
+const { policyFor } = require("../policy/index");
 
 async function store(req, res, next) {
   try {
     const payload = req.body;
+    const policy = policyFor(req.user);
+    if (!policy.can("create", "Tag")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memliki akses",
+      });
+    }
 
     const tag = new Tag(payload);
     await tag.save();
@@ -22,6 +30,14 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    const policy = policyFor(req.user);
+    if (!policy.can("update", "Tag")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memliki akses",
+      });
+    }
+
     const payload = req.body;
     const tag = await Tag.findByIdAndUpdate(req.params.id, payload, {
       new: true,
@@ -43,6 +59,14 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    const policy = policyFor(req.user);
+    if (!policy.can("delete", "Tag")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memliki akses",
+      });
+    }
+
     const tag = await Tag.findByIdAndDelete(req.params.id);
 
     return res.status(200).json(tag);
